+++ src/inputscope/conf.py
@@ -36,7 +36,7 @@ import sys
 
 """Program title, version number and version date."""
 Title = "InputScope"
-Version = "1.7.dev0"
+Version = "1.7.dev1"
 VersionDate = "23.07.2022"
 
 """TCP port of the web user interface."""