+++ src/inputscope/main.py
@@ -5,7 +5,7 @@ command-line echoer otherwise. Launches the event listener and web UI server.
 
 @author      Erki Suurjaak
 @created     05.05.2015
-@modified    07.07.2022
+@modified    23.07.2022
 """
 import calendar
 import datetime
@@ -99,6 +99,13 @@ class Model(threading.Thread):
         q = self.listenerqueue or self.initialqueue
         q.put("%s %s" % ("start" if on else "stop", category))
 
+
+    def set_config(self, name, value):
+        """Sets flag value in configuration and sends to listener."""
+        setattr(conf, name, value)
+        conf.save()
+        q = self.listenerqueue or self.initialqueue
+        q.put("configure %s %r" % (name, value))
 
     def session_action(self, action, session=None, arg=None):
         """Carries out session action."""
@@ -224,6 +231,7 @@ class MainApp(getattr(wx, "App", object)):
         on_session  = lambda k, s=None: functools.partial(self.OnSessionAction, k, session=s)
         on_clear    = lambda p, c, s=False: functools.partial(self.OnSessionAction, "clear", category=c) \
                                             if s else functools.partial(self.OnClearHistory, p, c)
+        on_toggle  = lambda n: functools.partial(self.OnToggleFlag, n)
 
         item_ui       = makeitem(menu, "&Open statistics")
         item_startup  = makeitem(menu, "Start with &Windows",  kind=wx.ITEM_CHECK) \
@@ -238,6 +246,7 @@ class MainApp(getattr(wx, "App", object)):
         item_scrolls  = makeitem(keyboardmenu, "Log mouse &scrolls",    kind=wx.ITEM_CHECK)
         item_keys     = makeitem(keyboardmenu, "Log individual &keys",  kind=wx.ITEM_CHECK)
         item_combos   = makeitem(keyboardmenu, "Log key &combinations", kind=wx.ITEM_CHECK)
+        item_sticky   = makeitem(keyboardmenu, "&Sticky long keypress", kind=wx.ITEM_CHECK)
 
         lastsession = self.model.sessions[0] if self.model.sessions else None
         activesession = lastsession if lastsession and not lastsession["end"] else None
@@ -293,6 +302,8 @@ class MainApp(getattr(wx, "App", object)):
         mousemenu.Append(item_scrolls)
         keyboardmenu.Append(item_keys)
         keyboardmenu.Append(item_combos)
+        keyboardmenu.AppendSeparator()
+        keyboardmenu.Append(item_sticky)
 
         menu.Append(item_ui)
         menu.Append(item_startup) if item_startup else None
@@ -319,6 +330,7 @@ class MainApp(getattr(wx, "App", object)):
         item_keyboard.Check(conf.KeyboardEnabled)
         item_keys    .Check(conf.KeyboardEnabled and conf.KeyboardKeysEnabled)
         item_combos  .Check(conf.KeyboardEnabled and conf.KeyboardCombosEnabled)
+        item_sticky  .Check(conf.KeyboardStickyEnabled)
         item_console .Check(self.frame_console.Shown)
         item_session_stop.Enable(bool(activename))
 
@@ -336,6 +348,7 @@ class MainApp(getattr(wx, "App", object)):
         menu.Bind(wx.EVT_MENU, on_session("stop"),      item_session_stop)
         menu.Bind(wx.EVT_MENU, self.OnToggleConsole,    item_console)
         menu.Bind(wx.EVT_MENU, self.OnClose,            item_exit)
+        menu.Bind(wx.EVT_MENU, on_toggle("KeyboardStickyEnabled"), item_sticky)
         self.trayicon.PopupMenu(menu)
 
 
@@ -398,6 +411,8 @@ class MainApp(getattr(wx, "App", object)):
     def OnToggleCategory(self, category, event):
         self.model.toggle(category)
 
+    def OnToggleFlag(self, name, event):
+        self.model.set_config(name, not getattr(conf, name))
 
     def OnSessionAction(self, action, event, session=None, category=None):
         arg = None